+++ src/components/checkbox/checkbox.jsx
@@ -1,14 +1,18 @@
 import React from "react";
 import "./checkbox.less";
+import "./switch.less";
 
 export default class Checkbox extends React.Component {
     static defaultProps = {
         checked: false,
+        mode: "checkbox",
 
         caption: "",
         variant: 1,
+        transition: true,
 
-        clsCheckbox: "",
+        cls: "",
+        className: "",
         clsCheck: "",
         clsCaption: "",
 
@@ -17,52 +21,44 @@ export default class Checkbox extends React.Component {
         onUnCheck: ()=>{}
     };
 
-    static getDerivedStateFromProps(props, state){
-        if (props.checked !== state.initChecked) {
-            return {
-                checked: props.checked,
-                initChecked: props.checked
-            }
-        }
-        return null;
-    }
-
-    constructor(props){
-        super(props);
-        this.state = {
-            checked: props.checked,
-            initChecked: props.checked
-        };
-
-        this.onChangeHandler = this.onChangeHandler.bind(this);
-    }
-
-    onChangeHandler(e){
+    onChangeHandler = (e) => {
         const state = e.target.checked;
-        this.setState({
-            checked: state
-        });
         this.props.onChange(e);
         this.props[state ? 'onCheck' : 'onUnCheck'](e);
-    }
+    };
 
     render(){
         const {
+            transition,
+            mode,
             checked,
             caption,
             variant,
+            cls, className,
             clsCaption,
-            clsCheckbox,
             clsCheck,
+            onChange,
+            onCheck,
+            onUnCheck,
             ...input
         } = this.props;
 
+        const checkboxMode = mode === "switch" ? `switch${(variant === 2 ? '-material' : '')}` : `checkbox ${(variant === 2 ? 'style2' : '')}`;
+
         return (
-            <label className={'checkbox' + clsCheckbox + ' ' + (variant === 2 ? 'style2' : '') + ' transition-on'}>
-                <input type="checkbox" { ...input } checked={this.state.checked} onChange={this.onChangeHandler}/>
+            <label className={`${checkboxMode} ${transition ? "transition-on" : ""} ${cls} ${className}`}>
+                <input type="checkbox" { ...input } defaultChecked={checked} onChange={this.onChangeHandler}/>
                 <span className={ 'check ' + clsCheck } />
                 <span className={ 'caption ' + clsCaption }>{caption}</span>
             </label>
         )
     }
 }
+
+const Switch = ({mode = "switch", ...rest}) => {
+    return (
+        <Checkbox mode={mode} {...rest}/>
+    )
+};
+
+export {Checkbox, Switch};