+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 + [x] Table: add props `emptyTitle`, `clsEmptyTitle`
 + [x] Button: add `overflow`
 + [x] Docs: fix for `IE11`
++ [x] Checkbox: optimize source
++ [x] Switch: optimize source
 
 ### 0.1.1
 + [x] Container: minor updates for using additional props